feat: added useData custom hook
