+++ src/hooks/useData.tsx
@@ -1,0 +1,7 @@
+import { DataContext } from 'contexts/DataContext/DataContext'
+import { useContext } from 'react'
+
+export const useData = () => {
+  const data = useContext(DataContext)
+  return data
+}
