+++ src/pages/dashboard/index.tsx
@@ -1,4 +1,3 @@
-import { useContext } from 'react'
 import { NextPage } from 'next'
 import Head from 'next/head'
 import dynamic from 'next/dynamic'
@@ -7,8 +6,7 @@ import { useRouter } from 'next/router'
 import { Container, Title } from '../../components/Layout/Base'
 import { ModalButton as Button } from '../../components/Layout/Buttons'
 import { ContainerRouter } from '../../styles/routeStyles'
-
-import { ModalContext } from '../../contexts/ModalContext/DataContext'
+import { useData } from 'hooks/useData'
 
 const Footer = dynamic(() => import('../../components/Footer'))
 
@@ -18,7 +16,7 @@ const Body = dynamic(
 
 const Dashboard: NextPage = () => {
   const router = useRouter()
-  const { data } = useContext(ModalContext)
+  const { data } = useData()
 
   return (
     <>
